+++ src/lua/CommunityBalanceMod/BalanceValueChanges/BalanceMisc.lua
@@ -1,0 +1,4 @@
+
+kPlayerEnergyPerEnergize  = 7.5
+
+kSecondInitialInfantryPortalMinPlayerCount = 7